Move ExternalLibraryName/ExternalLibraryNames
