+++ src/sagas/index.ts
@@ -8,8 +8,9 @@ import { call, put, race, select, take, takeEvery } from 'redux-saga/effects'
 import * as actions from '../actions'
 import * as actionTypes from '../actions/actionTypes'
 import { WorkspaceLocation } from '../actions/workspaces'
+import { ExternalLibraryNames } from '../components/assessment/assessmentShape'
 import { mockBackendSaga } from '../mocks/backend'
-import { externalLibraries, ExternalLibraryNames } from '../reducers/externalLibraries'
+import { externalLibraries } from '../reducers/externalLibraries'
 import { defaultEditorValue, IState } from '../reducers/states'
 import { IVLE_KEY, USE_BACKEND } from '../utils/constants'
 import { showSuccessMessage, showWarningMessage } from '../utils/notification'
